+++ beam_postgres/io/__init__.py
@@ -1,1 +1,6 @@
-from .postgres import ReadFromPostgres, WriteToPostgres
+from .postgres import ReadAllFromPostgres, WriteToPostgres
+from .retry import (
+    AlwaysRetryRowStrategy,
+    RetryRowOnTransientErrorStrategy,
+    RetryRowStrategy,
+)